+++ original/Code/WWAudio/SoundSceneObj.h
@@ -88,7 +88,7 @@ class SoundSceneObjClass : public MultiListObjectClass, public PersistClass, pub
 		//////////////////////////////////////////////////////////////////////
 		//	Event handling
 		//////////////////////////////////////////////////////////////////////		
-		virtual void			On_Event (AudioCallbackClass::EVENTS event, uint32_t param1 = 0, uint32_t param2 = 0);
+		virtual void			On_Event (AudioCallbackClass::EVENTS event, uintptr_t param1 = 0, uintptr_t param2 = 0);
 		virtual void			Register_Callback (AudioCallbackClass::EVENTS events, AudioCallbackClass *callback);
 		virtual void			Remove_Callback (void);
 
@@ -190,8 +190,8 @@ __inline void
 SoundSceneObjClass::On_Event
 (
 	AudioCallbackClass::EVENTS	event,
-	uint32_t							param1,
-	uint32_t							param2
+	uintptr_t							param1,
+	uintptr_t							param2
 )
 {
 	if ((m_pCallback != NULL) && (m_RegisteredEvents & event)) {
